fix: 본인이 작성하지 않은 게시글 상세 조회되지 않는 문제 해결
본인이 작성하지 않은 게시글에는수정/삭제 버튼이 보이지 않기 때문에 null 요소에 이벤트리스너를 할당하는 것이 된다는 점이 원인

+++ board/src/main/resources/static/js/article/article.js
@@ -1,10 +1,8 @@
 window.onload = function () {
 
     const urlStr = window.location.href;
-    const url = new URL(urlStr);
-    const urlPathName = url.pathname;
-    const lastStr = urlPathName.substring(urlPathName.length - 1);
-    const id = parseInt(lastStr);
+    const pathVariable = urlStr.split('/').pop();
+    const id = parseInt(pathVariable);
     console.log("id: ", id);
     console.log("typeof id: ", typeof id);
 
@@ -35,23 +33,26 @@ window.onload = function () {
         });
     })
 
-    document.querySelector('#article-delete-button').addEventListener('click', function(e) {
-        if (confirm('게시글을 삭제하시겠습니까?')) {
-            fetch(`/api/article/${id}`, {
-                method: 'DELETE',
-                headers: {
-                    'Content-Type': 'application/json',
-                }
-            }).then(response => {
-                console.log(response);
-                if (response.ok) {
-                    alert('게시물이 삭제되었습니다.');
-                    window.location.href = '/dashboard';
-                }
-            }).catch(error => {
-                console.log('Error: ', error);
-            });
-        }
+    const deleteButton = document.querySelector('#article-delete-button');
+    if (deleteButton !== null) {
+        document.querySelector('#article-delete-button').addEventListener('click', function(e) {
+            if (confirm('게시글을 삭제하시겠습니까?')) {
+                fetch(`/api/article/${id}`, {
+                    method: 'DELETE',
+                    headers: {
+                        'Content-Type': 'application/json',
+                    }
+                }).then(response => {
+                    console.log(response);
+                    if (response.ok) {
+                        alert('게시물이 삭제되었습니다.');
+                        window.location.href = '/dashboard';
+                    }
+                }).catch(error => {
+                    console.log('Error: ', error);
+                });
+            }
+        });
+    }
 
-    });
 };
